add encoding package
